FGFFGFG

+++ frontend/lib/screens/home/inventory_screen_backup.dart
@@ -3052,7 +3052,7 @@ class _ProductDialogState extends State<_ProductDialog> {
         'description': _descriptionController.text.trim(),
         'price': double.parse(_priceController.text),
         'cost_price': double.parse(_costController.text),
-        'stock_quantity': int.parse(_stockController.text),
+        'stock_quantity': double.parse(_stockController.text),
         'category_id': _selectedCategoryId,
         'sku': _skuController.text.trim(),
         'low_stock_threshold': 10, // Default value
